backend/requirements.txt
fastapi
uvicorn
tensorflow==2.13.0
pillow
numpy
scikit-learn
pandas
matplotlib
